psdev-296: Fix deployment with LDAP supported.
Fix LDAP with a hack workaround described in
https://github.com/jupyterhub/ldapauthenticator/issues/128.
The hub dockerfile applies a patch to the ldapathenticator py module.

This is based on upstream version 0.9-b609a67 helm chart.
This was obtained from a tarball so a bit disconnected from upstream. However, it is working so checking in.

DCO 1.1 Signed-off-by: David Ohlemacher <dohlemacher@infinidat.com>